Add docs, start cleaning up
