+++ fmm/src/charge.rs
@@ -1,3 +1,4 @@
+//! Helper functions to handle charge data.
 use std::collections::HashMap;
 
 use cauchy::Scalar;
@@ -5,6 +6,11 @@ use num::Float;
 
 use crate::types::{Charge, ChargeDict, GlobalIdx};
 
+/// Zip together ordered list of global indices with their associated charges in a dictionary.
+/// 
+/// # Arguments
+/// * `global_idxs` - Unique global index associated with charge.
+/// * `charges` - The charge associated with each unique global index.
 pub fn build_charge_dict<T>(global_idxs: &[GlobalIdx], charges: &[Charge<T>]) -> ChargeDict<T>
 where
     T: Float + Scalar + Default,
