+++ fmm/src/types.rs
@@ -1,3 +1,4 @@
+//! Data structures FMM data and metadata.
 use std::collections::HashMap;
 
 use bempp_traits::{field::FieldTranslationData, fmm::Fmm, kernel::Kernel, tree::Tree};
@@ -218,6 +219,110 @@ impl<T> Default for SendPtr<T> {
     fn default() -> Self {
         SendPtr {
             raw: std::ptr::null(),
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+
+
+    use bempp_field::types::FftFieldTranslationKiFmm;
+    use itertools::Itertools;
+    use bempp_tree::{implementations::helpers::points_fixture, types::single_node::SingleNodeTree};
+    use bempp_traits::tree::Tree;
+    use bempp_traits::field::FieldTranslationData;
+    use bempp_traits::fmm::Fmm;
+    use bempp_kernel::laplace_3d::Laplace3dKernel;
+    use rlst::dense::RawAccess;
+    use crate::{charge::build_charge_dict, types::{FmmDataUniform, KiFmmLinear}};
+
+    #[test]
+    fn test_fmm_data_uniform() {
+        let npoints = 10000;
+        let points = points_fixture::<f64>(npoints, None, None);
+        let global_idxs = (0..npoints).collect_vec();
+        let charges = vec![1.0; npoints];
+
+        let order = 8;
+        let alpha_inner = 1.05;
+        let alpha_outer = 2.95;
+        let ncrit = 150;
+        let depth = 3;
+
+        // Uniform trees
+        {
+            let adaptive = false;
+            let kernel = Laplace3dKernel::default();
+
+            let tree = SingleNodeTree::new(
+                points.data(),
+                adaptive,
+                Some(ncrit),
+                Some(depth),
+                &global_idxs[..],
+                false,
+            );
+
+            let m2l_data_fft = FftFieldTranslationKiFmm::new(
+                kernel.clone(),
+                order,
+                *tree.get_domain(),
+                alpha_inner,
+            );
+
+            let fmm = KiFmmLinear::new(order, alpha_inner, alpha_outer, kernel, tree, m2l_data_fft);
+            // Form charge dict, matching charges with their associated global indices
+            let charge_dict = build_charge_dict(&global_idxs[..], &charges[..]);
+
+            let datatree = FmmDataUniform::new(fmm, &charge_dict).unwrap();
+
+            let ncoeffs = datatree.fmm.m2l.ncoeffs(order);
+            let nleaves = datatree.fmm.tree().get_all_leaves_set().len();
+            let nkeys = datatree.fmm.tree().get_all_keys_set().len();
+
+            // Test that the number of of coefficients is being correctly assigned
+            assert_eq!(datatree.multipoles.len(), ncoeffs * nkeys);
+            assert_eq!(datatree.leaf_multipoles.len(), nleaves);
+
+            // Test that leaf indices are being mapped correctly to leaf multipoles
+            let idx = 0;
+            let leaf_key = &datatree.fmm.tree().get_all_leaves().unwrap()[idx];
+            let &leaf_idx = datatree.fmm.tree().get_index(leaf_key).unwrap();
+            unsafe {
+                let result = datatree.multipoles.as_ptr().add(leaf_idx * ncoeffs);
+                let expected =
+                    datatree.multipoles[leaf_idx * ncoeffs..(leaf_idx + 1) * ncoeffs].as_ptr();
+                assert_eq!(result, expected);
+
+                let result = datatree.locals.as_ptr().add(leaf_idx * ncoeffs);
+                let expected =
+                    datatree.locals[leaf_idx * ncoeffs..(leaf_idx + 1) * ncoeffs].as_ptr();
+                assert_eq!(result, expected);
+            }
+
+            // Test that level expansion information is referring to correct memory, and is in correct shape
+            assert_eq!(
+                datatree.level_multipoles.len() as u64,
+                datatree.fmm.tree().get_depth() + 1
+            );
+            assert_eq!(
+                datatree.level_locals.len() as u64,
+                datatree.fmm.tree().get_depth() + 1
+            );
+
+            assert_eq!(
+                datatree.level_multipoles[(datatree.fmm.tree().get_depth()) as usize].len(),
+                nleaves
+            );
+            assert_eq!(
+                datatree.level_locals[(datatree.fmm.tree().get_depth()) as usize].len(),
+                nleaves
+            );
+
+            // Test that points are being assigned correctly
+            assert_eq!(datatree.potentials_send_pointers.len(), nleaves);
+            assert_eq!(datatree.potentials.len(), npoints);
         }
     }
 }
