+++ traits/src/field.rs
@@ -55,5 +55,8 @@ where
     /// Particle to local translations, applies to leaf boxes when a source box is within
     /// the far field of a target box, but is too large for the multipole expansion to converge
     /// at the target, so instead its contribution is computed directly.
+    /// 
+    /// # Arguments
+    /// * `level` - The level of the tree at which a field translation is being applied.
     fn p2l(&self, level: u64);
 }
