+++ tree/src/implementations/helpers.rs
@@ -16,8 +16,8 @@ pub type PointsMat<T> = Matrix<T, BaseMatrix<T, VectorContainer<T>, Dynamic>, Dy
 ///
 /// # Arguments
 /// * `npoints` - The number of points to sample.
-/// # `min` - The minumum coordinate value along each axis.
-/// # `max` - The maximum coordinate value along each axis.
+/// * `min` - The minumum coordinate value along each axis.
+/// * `max` - The maximum coordinate value along each axis.
 pub fn points_fixture<T: Float + Scalar + rand::distributions::uniform::SampleUniform>(
     npoints: usize,
     min: Option<T>,
@@ -48,8 +48,8 @@ pub fn points_fixture<T: Float + Scalar + rand::distributions::uniform::SampleUn
 ///
 /// # Arguments
 /// * `npoints` - The number of points to sample.
-/// # `min` - The minumum coordinate value along each axis.
-/// # `max` - The maximum coordinate value along each axis.
+/// * `min` - The minumum coordinate value along each axis.
+/// * `max` - The maximum coordinate value along each axis.
 pub fn points_fixture_sphere<T: Scalar + rand::distributions::uniform::SampleUniform>(
     npoints: usize,
 ) -> PointsMat<T> {
